Load global scalars in setup so they don't get reloaded in loops
