+++ crates/cubecl-spirv/src/lookups.rs
@@ -1,6 +1,6 @@
 use std::collections::VecDeque;
 
-use cubecl_core::ir::KernelDefinition;
+use cubecl_core::ir::{self, KernelDefinition};
 use cubecl_opt::NodeIndex;
 use hashbrown::{HashMap, HashSet};
 use rspirv::spirv::{BuiltIn, CooperativeMatrixLayout, CooperativeMatrixUse, StorageClass, Word};
@@ -26,6 +26,7 @@ pub struct LookupTables {
 
     pub used_builtins: HashMap<BuiltIn, (Word, Item)>,
 
+    pub scalars: HashMap<(u16, ir::Elem), Word>,
     pub globals: HashMap<Globals, Word>,
     pub array_types: HashMap<Word, Word>,
     pub constants: HashMap<(ConstVal, Item), Word>,
@@ -235,6 +236,30 @@ impl<T: SpirvTarget> SpirvCompiler<T> {
             let word = self.label(block);
             self.state.end_labels.insert(block, word);
             word
+        }
+    }
+
+    pub fn global_scalar(&mut self, id: u16, elem: ir::Elem) -> Variable {
+        if let Some(existing) = self.state.scalars.get(&(id, elem)).copied() {
+            let item = self.compile_item(ir::Item::new(elem));
+            Variable::GlobalScalar(existing, item.elem())
+        } else {
+            let current_block = self.selected_block();
+            let setup = self.setup_block;
+            self.select_block(Some(setup)).unwrap();
+            let arr_id = self.state.named[&format!("scalars_{elem}")];
+            let item = self.compile_item(ir::Item::new(elem));
+            let arr = Variable::GlobalInputArray(arr_id, item.clone());
+            let const_id = self.const_u32(id as u32);
+            let index =
+                Variable::ConstantScalar(const_id, (id as u32).into(), Elem::Int(32, false));
+            let read_id = self.id();
+            let var = Variable::GlobalScalar(read_id, item.elem());
+            self.debug_name(read_id, format!("scalars<{elem}>({id})"));
+            self.read_indexed_unchecked(&var, &arr, &index);
+            self.select_block(current_block).unwrap();
+            self.state.scalars.insert((id, elem), read_id);
+            var
         }
     }
 }
